Fix mismatch between declaration and definition of __go_alloc and mlookup.  Make sure %zu sees size_t.  Fixes issue 356.
git-svn-id: svn+ssh://gcc.gnu.org/svn/gcc/branches/gccgo@155451 138bc75d-0d04-0410-961f-82ee72b054a4

+++ libgo/runtime/malloc.c
@@ -315,7 +315,7 @@ stackalloc(uint32 n)
 		if(stacks.size == 0)
 			FixAlloc_Init(&stacks, n, SysAlloc, nil, nil);
 		if(stacks.size != n) {
-			printf("stackalloc: in malloc, size=%zu want %d", stacks.size, n);
+			printf("stackalloc: in malloc, size=%zu want %d", (size_t)stacks.size, n);
 			throw("stackalloc");
 		}
 		v = FixAlloc_Alloc(&stacks);
